+++ src/app/list.ts
@@ -3,7 +3,7 @@ import { Table } from 'console-table-printer';
 import { SupabaseClient } from '@supabase/supabase-js';
 import * as p from '@clack/prompts';
 import { Database } from 'types/supabase.types';
-import { OptionsBase, createSupabaseClient, findSavedKey, formatError, getHumanDate, verifyUser } from '../utils';
+import { OptionsBase, createSupabaseClient, findSavedKey, getHumanDate, verifyUser } from '../utils';
 import { checkLatest } from '../api/update';
 
 const displayApp = (data: Database['public']['Tables']['apps']['Row'][]) => {
@@ -16,6 +16,7 @@ const displayApp = (data: Database['public']['Tables']['apps']['Row'][]) => {
   data.reverse().forEach(row => {
     t.addRow({
       Name: row.name,
+      id: row.app_id,
       Created: getHumanDate(row.created_at)
     });
   });