+++ src/bundle/compatibility.ts
@@ -5,7 +5,8 @@ import { createSupabaseClient, findSavedKey, getConfig, verifyUser, checkCompati
 import { checkAppExistsAndHasPermissionErr } from '../api/app';
 
 interface Options extends OptionsBase {
-    channel?: string
+    channel?: string,
+    text?: boolean,
 }
 
 export const checkCompatibilityCommand = async (appId: string, options: Options) => {
@@ -45,7 +46,7 @@ export const checkCompatibilityCommand = async (appId: string, options: Options)
     // const nativePackages = Array.from(hashedLocalDependencies, ([name, value]) => ({ name, version: value.version }))
     // await supabase.from('app_versions').update({ native_packages: nativePackages }).eq('id', '9654')
 
-    const { finalCompatibility } = await checkCompatibility(supabase, channel)
+    const { finalCompatibility } = await checkCompatibility(supabase, appId, channel)
 
 
     const t = new Table({
@@ -53,6 +54,10 @@ export const checkCompatibilityCommand = async (appId: string, options: Options)
         charLength: { "❌": 2, "✅": 2 },
     });
     
+
+    const yesSymbol = options.text ? 'Yes' : '✅'
+    const noSymbol = options.text ? 'No' : '❌'
+
     finalCompatibility.forEach((data) => {
         const { name, localVersion, remoteVersion } = data
 
@@ -60,7 +65,7 @@ export const checkCompatibilityCommand = async (appId: string, options: Options)
             Package: name,
             'Local version': localVersion ?? 'None',
             'Remote version': remoteVersion ?? 'None',
-            Compatible: remoteVersion === localVersion ? '✅' : '❌',
+            Compatible: remoteVersion === localVersion ? yesSymbol : noSymbol,
         });
     })